move eventHandlers to contractEventSubscribers
