+++ buyer-api/src/blockchain/dataOrder.js
@@ -1,10 +1,5 @@
-import apicache from 'apicache';
+import { DataExchange, toDate } from './contracts';
 import { toWib } from '../utils/wibson-lib/coin';
-import { DataExchange } from './contracts';
-import { toDate } from './utils';
-import { contractEventListener } from './contractEventListener';
-import { getAccount } from '../services/signingService';
-import { dataOrders } from '../utils/stores';
 
 /**
  * @typedef DataOrder DataOrder information on the DataExchange
@@ -46,36 +41,3 @@ export async function fetchDataOrder(dxId) {
   }
   throw new Error('Invalid Id');
 }
-
-/**
- * @typedef DataOrderEventValues
- * @property {string} owner The buyer that created the DataOrder
- * @property {number} orderId The DataExchange id of the DataOrder
-
- * @callback dataOrderUpdapter Updates DataOrder in the store with data from the DataExchange
- * @param {DataOrderEventValues} eventValues The values emmited by the DataExchange event
-
- * @function createDataOrderUpdapter Creates a dataOrderUpdapter with the given status
- * @param {string} status The status to be set by the updater
- * @returns {dataOrderUpdapter}
- */
-const createDataOrderUpdapter = status => async ({ owner, orderId: dxId }) => {
-  const { address } = await getAccount();
-  if (address.toLowerCase() === owner.toLowerCase()) {
-    const { buyer, ...chainOrder } = await fetchDataOrder(dxId);
-    const id = chainOrder.buyerUrl.match(/\/orders\/(.+)\/offchain-data/)[1];
-    const storedOrder = await dataOrders.fetch(id);
-    await dataOrders.store(id, {
-      ...storedOrder,
-      ...chainOrder,
-      dxId,
-      status,
-    });
-    apicache.clear('/orders/*');
-  }
-};
-
-contractEventListener
-  .addContract(DataExchange)
-  .on('DataOrderCreated', createDataOrderUpdapter('created'))
-  .on('DataOrderClosed', createDataOrderUpdapter('closed'));
