+++ buyer-api/src/server.js
@@ -2,7 +2,7 @@ import 'babel-polyfill';
 import app from './app';
 import config from '../config';
 import { logger } from './utils';
-import { contractEventListener } from './blockchain/contractEventListener';
+import { contractEventListener } from './blockchain/contractEventSubscribers';
 
 const server = () => {
   const { port, host, env } = config;
